+++ examples/guard_custom_error.rs
@@ -34,13 +34,13 @@ statemachine! {
 pub struct Context;
 
 impl StateMachineContext for Context {
-    type GuardError = GuardError; // Guard1 has access to the data from Event1
+    type Error = GuardError; // Guard1 has access to the data from Event1
     fn guard1(&self, _event_data: &MyEventData) -> Result<bool, GuardError> {
         Err(GuardError::Custom)
     }
 
     // Action1 has access to the data from Event1, and need to return the state data for State2
-    fn action1(&mut self, _event_data: MyEventData) -> Result<MyStateData, Self::GuardError> {
+    fn action1(&mut self, _event_data: MyEventData) -> Result<MyStateData, Self::Error> {
         todo!()
     }
 
@@ -50,7 +50,7 @@ impl StateMachineContext for Context {
     }
 
     // Action2 has access to the data from State2
-    fn action2(&mut self, _state_data: &MyStateData) -> Result<(), Self::GuardError> {
+    fn action2(&mut self, _state_data: &MyStateData) -> Result<(), Self::Error> {
         todo!()
     }
 }